+++ libs/common/src/entities/index.ts
@@ -2,4 +2,5 @@ export * from './activities.entity';
 export * from './crops.entity';
 export * from './farms.entity';
 export * from './harvest.entity';
+export * from './outbox.entity';
 export * from './user.entity';